refactor: Implement Quick Access section in Home component for streamlined user navigation
